python/requirements.txt
openai-whisper
mediapipe
opencv-python
moviepy
ffmpeg-python
numpy
torch
yt-dlp
